Finalize QR code with information like price and description
