+++ app/src/main/res/layout/activity_detail_buy.xml
@@ -45,7 +45,7 @@
                     android:layout_marginBottom="10dp"
                     android:text="Prix"
                     android:textColor="@android:color/holo_red_light"
-                    android:textSize="100sp" />
+                    android:textSize="50sp" />
             </android.support.v7.widget.CardView>
 
             <TextView
@@ -53,7 +53,6 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="40dp"
-                android:lines="4"
                 android:hint="Detail sur le produit"
                 android:textColor="@android:color/black"
                 android:textSize="18sp" />
